+++ omni-icon.php
@@ -6,7 +6,7 @@
  * Plugin URI:          https://github.com/nabasa-dev/omni-icon
  * Description:         A modern SVG icon library for WordPress with support for custom uploads and 200,000+ Iconify icons across block editor, page builders, and themes.
  * Text Domain:         omni-icon
- * Version:             1.0.13
+ * Version:             1.0.14
  * Requires at least:   6.0
  * Requires PHP:        8.1
  * Author:              Omni Icon