+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Increase webcomponent concurrent request limit from 4 to 16 for faster icon loading
+
 ## [1.0.13] - 2026-01-13
 
 ### Fixed